refactor: only export fonts
